Modified HoM.ned parameter, max_buffer to type int from double. Also caster par(Photon_detection_per_sec) to int when copying to an integer variable.

+++ quisp/modules/HoM_Controller.cc
@@ -214,7 +214,7 @@ void HoM_Controller::checkNeighborBuffer(bool receiver){
 
 void HoM_Controller::updateIDE_Parameter(bool receiver){
     try{
-        photon_detection_per_sec = par("photon_detection_per_sec");
+        photon_detection_per_sec = (int)par("photon_detection_per_sec");
         if(photon_detection_per_sec <= 0){
             error("Photon detection per sec for HoM must be more than 0.");
         }
